+++ java-snapshot-testing-plugin-jackson/src/test/java/au/com/origin/snapshots/jackson/ReflectionUtilities.java
@@ -1,4 +1,4 @@
-package au.com.origin.snapshots;
+package au.com.origin.snapshots.jackson;
 
 import au.com.origin.snapshots.exceptions.SnapshotMatchException;
