+++ java-snapshot-testing-plugin-jackson/src/test/java/au/com/origin/snapshots/jackson/serializers/docs/CustomSerializerTest.java
@@ -1,4 +1,4 @@
-package au.com.origin.snapshots.docs;
+package au.com.origin.snapshots.jackson.serializers.docs;
 
 import au.com.origin.snapshots.Expect;
 import au.com.origin.snapshots.PropertyResolvingSnapshotConfig;
