+++ java-snapshot-testing-plugin-jackson/src/test/java/au/com/origin/snapshots/jackson/serializers/docs/JsonAssertReporter.java
@@ -1,4 +1,4 @@
-package au.com.origin.snapshots.docs;
+package au.com.origin.snapshots.jackson.serializers.docs;
 
 import au.com.origin.snapshots.reporters.SnapshotReporter;
 import au.com.origin.snapshots.serializers.SerializerType;
